+++ packages/image_picker/image_picker_for_web/lib/image_picker_for_web.dart
@@ -42,70 +42,6 @@ class ImagePickerPlugin extends ImagePickerPlatform {
     ImagePickerPlatform.instance = ImagePickerPlugin();
   }
 
-  /// Returns a [PickedFile] with the image that was picked.
-  ///
-  /// The `source` argument controls where the image comes from. This can
-  /// be either [ImageSource.camera] or [ImageSource.gallery].
-  ///
-  /// Note that the `maxWidth`, `maxHeight` and `imageQuality` arguments are not supported on the web. If any of these arguments is supplied, it'll be silently ignored by the web version of the plugin.
-  ///
-  /// Use `preferredCameraDevice` to specify the camera to use when the `source` is [ImageSource.camera].
-  /// The `preferredCameraDevice` is ignored when `source` is [ImageSource.gallery]. It is also ignored if the chosen camera is not supported on the device.
-  /// Defaults to [CameraDevice.rear].
-  ///
-  /// If no images were picked, the return value is null.
-  @override
-  Future<PickedFile> pickImage({
-    required ImageSource source,
-    double? maxWidth,
-    double? maxHeight,
-    int? imageQuality,
-    CameraDevice preferredCameraDevice = CameraDevice.rear,
-  }) {
-    final String? capture =
-        computeCaptureAttribute(source, preferredCameraDevice);
-    return pickFile(accept: _kAcceptImageMimeType, capture: capture);
-  }
-
-  /// Returns a [PickedFile] containing the video that was picked.
-  ///
-  /// The [source] argument controls where the video comes from. This can
-  /// be either [ImageSource.camera] or [ImageSource.gallery].
-  ///
-  /// Note that the `maxDuration` argument is not supported on the web. If the argument is supplied, it'll be silently ignored by the web version of the plugin.
-  ///
-  /// Use `preferredCameraDevice` to specify the camera to use when the `source` is [ImageSource.camera].
-  /// The `preferredCameraDevice` is ignored when `source` is [ImageSource.gallery]. It is also ignored if the chosen camera is not supported on the device.
-  /// Defaults to [CameraDevice.rear].
-  ///
-  /// If no images were picked, the return value is null.
-  @override
-  Future<PickedFile> pickVideo({
-    required ImageSource source,
-    CameraDevice preferredCameraDevice = CameraDevice.rear,
-    Duration? maxDuration,
-  }) {
-    final String? capture =
-        computeCaptureAttribute(source, preferredCameraDevice);
-    return pickFile(accept: _kAcceptVideoMimeType, capture: capture);
-  }
-
-  /// Injects a file input with the specified accept+capture attributes, and
-  /// returns the PickedFile that the user selected locally.
-  ///
-  /// `capture` is only supported in mobile browsers.
-  /// See https://caniuse.com/#feat=html-media-capture
-  @visibleForTesting
-  Future<PickedFile> pickFile({
-    String? accept,
-    String? capture,
-  }) {
-    final html.FileUploadInputElement input =
-        createInputElement(accept, capture) as html.FileUploadInputElement;
-    _injectAndActivate(input);
-    return _getSelectedFile(input);
-  }
-
   /// Returns an [XFile] with the image that was picked.
   ///
   /// The `source` argument controls where the image comes from. This can
@@ -119,7 +55,7 @@ class ImagePickerPlugin extends ImagePickerPlatform {
   ///
   /// If no images were picked, the return value is null.
   @override
-  Future<XFile> getImage({
+  Future<XFile?> getImage({
     required ImageSource source,
     double? maxWidth,
     double? maxHeight,
@@ -132,12 +68,14 @@ class ImagePickerPlugin extends ImagePickerPlatform {
       accept: _kAcceptImageMimeType,
       capture: capture,
     );
-    return _imageResizer.resizeImageIfNeeded(
-      files.first,
-      maxWidth,
-      maxHeight,
-      imageQuality,
-    );
+    return files.isEmpty
+        ? null
+        : _imageResizer.resizeImageIfNeeded(
+            files.first,
+            maxWidth,
+            maxHeight,
+            imageQuality,
+          );
   }
 
   /// Returns an [XFile] containing the video that was picked.
@@ -153,7 +91,7 @@ class ImagePickerPlugin extends ImagePickerPlatform {
   ///
   /// If no images were picked, the return value is null.
   @override
-  Future<XFile> getVideo({
+  Future<XFile?> getVideo({
     required ImageSource source,
     CameraDevice preferredCameraDevice = CameraDevice.rear,
     Duration? maxDuration,
@@ -164,7 +102,7 @@ class ImagePickerPlugin extends ImagePickerPlatform {
       accept: _kAcceptVideoMimeType,
       capture: capture,
     );
-    return files.first;
+    return files.isEmpty ? null : files.first;
   }
 
   /// Injects a file input, and returns a list of XFile images that the user selected locally.
@@ -267,29 +205,6 @@ class ImagePickerPlugin extends ImagePickerPlatform {
     return input == null ? null : _getFilesFromInput(input);
   }
 
-  /// Monitors an <input type="file"> and returns the selected file.
-  Future<PickedFile> _getSelectedFile(html.FileUploadInputElement input) {
-    final Completer<PickedFile> completer = Completer<PickedFile>();
-    // Observe the input until we can return something
-    input.onChange.first.then((html.Event event) {
-      final List<html.File>? files = _handleOnChangeEvent(event);
-      if (!completer.isCompleted && files != null) {
-        completer.complete(PickedFile(
-          html.Url.createObjectUrl(files.first),
-        ));
-      }
-    });
-    input.onError.first.then((html.Event event) {
-      if (!completer.isCompleted) {
-        completer.completeError(event);
-      }
-    });
-    // Note that we don't bother detaching from these streams, since the
-    // "input" gets re-created in the DOM every time the user needs to
-    // pick a file.
-    return completer.future;
-  }
-
   /// Monitors an <input type="file"> and returns the selected file(s).
   Future<List<XFile>> _getSelectedXFiles(html.FileUploadInputElement input) {
     final Completer<List<XFile>> completer = Completer<List<XFile>>();
@@ -310,6 +225,11 @@ class ImagePickerPlugin extends ImagePickerPlatform {
         }).toList());
       }
     });
+
+    input.addEventListener('cancel', (html.Event _) {
+      completer.complete(<XFile>[]);
+    });
+
     input.onError.first.then((html.Event event) {
       if (!completer.isCompleted) {
         completer.completeError(event);
@@ -361,6 +281,7 @@ class ImagePickerPlugin extends ImagePickerPlatform {
   void _injectAndActivate(html.Element element) {
     _target.children.clear();
     _target.children.add(element);
+    // TODO(dit): Reimplement this with the showPicker() API, https://github.com/flutter/flutter/issues/130365
     element.click();
   }
 }